requirements.txt
fastapi
uvicorn[standard]
aio-pika
python-dotenv
pydantic